Title: 앞에서부터 삭제하기 2, Time: 343ms, Memory: 21MB, Status: Passed - Codetree

+++ 250501/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.java
@@ -27,7 +27,7 @@ public class Main {
             sum += answer[k];
 
             int min = pq.peek();
-            maxAvg = Math.max(maxAvg, ((sum - min) / (n - k)));
+            maxAvg = Math.max(maxAvg, ((float)(sum - min) / (n - k)));
             
         }
